+++ Assets/DeltaDNA/DDNA.cs
@@ -503,7 +503,7 @@ namespace DeltaDNA
         /// </summary>
         public string CollectURL {
             get { return this.collectURL; }
-            private set { this.collectURL = ValidateURL(value); }
+            private set { this.collectURL = Utils.FixURL(value); }
         }
 
         /// <summary>
@@ -511,7 +511,7 @@ namespace DeltaDNA
         /// </summary>
         public string EngageURL {
             get { return this.engageURL; }
-            private set { this.engageURL = ValidateURL(value); }
+            private set { this.engageURL = Utils.FixURL(value); }
         }
 
         /// <summary>
@@ -790,13 +790,6 @@ namespace DeltaDNA
             uri = uri.Replace("{env_key}", envKey);
             uri = uri.Replace("{hash}", hash);
             return uri;
-        }
-
-        private static string ValidateURL(string url) {
-            if (!url.ToLower().StartsWith("http://") && !url.ToLower().StartsWith("https://")) {
-                url = "https://" + url;
-            }
-            return url;
         }
 
         private static string GenerateHash(string data, string secret){